modularize and test tracing where should not store message
